refactor(improve): extract eligibility/candidate-selection helpers to improve/eligibility.ts (D1 PR-C)
Move the 15 free candidate-selection predicates (resolveImproveScope, dedupeRefs,
collectEligibleRefs, isEntryInScope/isEntryInWritableSource, memoryCleanupParentRef,
isLessonCandidate, isDistillCandidateRef, shouldDistillMemoryRef,
buildLatestFeedbackTsMap, buildLatestProposalTsMap, isSignalDeltaEligible,
shouldAnalyzeMemoryCleanup, buildUtilityMap, findAssetFilePath) into
improve/eligibility.ts. They are free functions (no orchestrator closure state),
so this is a behavior-preserving relocation that shrinks the god-module and makes
candidate selection independently testable. improve.ts 5,289 -> 4,797 LOC.
tsc 0 / biome 0 / unit 5274-0.
